Add NIF sext codec
This provides a subset of the sext encoding and decoding functionality
that should be enough for Basho's use in leveldb.  It supports binaries,
tuples, atoms and integers up to 64 bits.  On the encoding side, only
binary encoding has been implemented, but without bistring support.
The motivation for this change was speed.
Notice that NIF that may run for a long time (> 1ms) may confuse the
Erlang scheduler. From local measurements, it seems that it should not
be a worry for the size of Erlang terms that we are decoding. But that
may be revisited later.